pyproject.toml
[project]
name = "life_management_agency"
version = "0.1.0"
description = "Life Management Agency with Multiple AI Agents"
requires-python = ">=3.12"
dependencies = [
    "setuptools>=75.5.0",
    "python-dotenv",
    "agency-swarm",
    "pydantic",
    "tavily-python",
    "openai",
    "requests"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["life_management_agency"]